refactor(live-chat): 잘못된 패키지 위치 수정
