+++ src/main/java/org/codeit/sb06/team03/mopl/content/infra/in/ContentControllerAdvice.java
@@ -2,7 +2,7 @@ package org.codeit.sb06.team03.mopl.content.infra.in;
 
 import org.codeit.sb06.team03.mopl.common.error.ErrorResponse;
 import org.codeit.sb06.team03.mopl.content.domain.exception.InvalidCursorFormatException;
-import org.codeit.sb06.team03.mopl.liveChat.domain.exception.ContentNotFoundException;
+import org.codeit.sb06.team03.mopl.content.domain.exception.ContentNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
